Reload proposals on creation
